requirements.txt
kaggle
pyyaml
pandas
numpy
matplotlib
seaborn
scikit-learn
mlflow
catboost
fastapi
uvicorn
pydantic
gradio